Restructure `pep517_backend` package
