+++ pyproject.toml
@@ -13,7 +13,7 @@ requires = [
   "setuptools_scm_git_archive>=1.1",
 ]
 backend-path = ["bin"]  # requires 'Pip>=20' or 'pep517>=0.6.0'
-build-backend = "pep517_backend"
+build-backend = "pep517_backend.hooks"
 
 [tool.local.cythonize]
 # This attr can contain multiple globs
